+++ Entity/Job.php
@@ -20,7 +20,7 @@ class Job extends BaseJob
 {
     /**
      * @Grid\Column(sortable=true,order=1)
-     * @ORM\Column(type="bigint")
+     * @ORM\Column(name="id", type="bigint")
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="AUTO")
      */