+++ Entity/JobTiming.php
@@ -12,19 +12,19 @@ use Doctrine\ORM\Mapping as ORM;
 class JobTiming extends BaseJobTiming
 {
     /**
-     * @ORM\Column(type="bigint")
+     * @ORM\Column(name="id", type="bigint")
      * @ORM\Id
      * @ORM\GeneratedValue(strategy="AUTO")
      */
     protected $id;
 
     /**
-     * @ORM\Column(type="datetime")
+     * @ORM\Column(name="finished_at", type="datetime")
      */
     protected $finishedAt;
 
     /**
-     * @ORM\Column(type="integer")
+     * @ORM\Column(name="status", type="integer")
      */
     protected $status;